+++ services/autoscaling/src/simcore_service_autoscaling/modules/cluster_scaling/buffer_machines_pool_task.py
@@ -11,7 +11,7 @@ from ...core.settings import ApplicationSettings
 from ...utils.redis import create_lock_key_and_value
 from ..redis import get_redis_client
 from ._buffer_machines_pool_core import monitor_buffer_machines
-from .auto_scaling_mode_dynamic import DynamicAutoscaling
+from ._provider_dynamic import DynamicAutoscaling
 
 _TASK_NAME_BUFFER: Final[str] = "Autoscaling Buffer Machines Pool"
